Fix bug in how libremote connections are closed
When Libremote's close_connection() was invoked, it would send an 'exit'
command via send_command(). The problem was that the only implementation of
send_command() (i.e. the one in ShellRemote) was configured to expect an
'END_BPFD_OUTPUT' response from bpfd whenever any command was sent, but in the
case of the 'exit' command, bpfd never actually gave back any response. This
caused the connection to timeout instead of closing gracefully.

Signed-off-by: Jazel Canseco <jcanseco@google.com>